Handle cases when both layout and keyboard contain commas
A proper solution would be escaped commas in both name and layout, or a
different separator, or better yet, socket3 IPC with json.

But, since we only care about one (main) keyboard anyway, we can just
subtract it from the beginning of the event_data string, to get our
layout.

`Case` instead of `if grep -> sed...` etc for performance.